requirements.txt
Flask==2.3.3
requests==2.31.0
python-slugify==8.0.1
playsound==1.2.2
gunicorn==21.2.0